refactor(l5): move nats-runtime, auth, background-tracing from src/ into server/application (issue 864)
Cluster C4b of the L5 cut: three env-free runtime modules leave src/ for
server/application/, with their unit tests moved alongside them unmodified.

Moves (old -> new):

- src/nats-runtime.ts       -> server/application/nats-runtime.ts
- src/nats-runtime.test.ts  -> server/application/nats-runtime.test.ts
- src/auth.ts               -> server/application/auth.ts
- src/auth.test.ts          -> server/application/auth.test.ts
- src/background-tracing.ts      -> server/application/background-tracing.ts
- src/background-tracing.test.ts -> server/application/background-tracing.test.ts

New sibling, split by responsibility because nats-runtime.ts was 562 code
lines after prettier and the server/ standard is under 500:

- server/application/nats-envelope.ts -- the fleet envelope wire codec (the
  TS mirror of fleet_nats.envelope.Envelope plus its build/encode/decode
  helpers). nats-runtime.ts keeps the runtime-boundary and bridge-planning
  responsibility and drops to 400 lines; it re-exports the codec so every
  existing importer keeps its line.

Reaching the server/ standard, structure only, no behavior change:

- nats-envelope.ts: assertEnvelopeIdentity, decodeEnvelopeRecord,
  readEnvelopePayload, readEnvelopeVersion hoisted out of buildEnvelope and
  envelopeFromBytes.
- nats-runtime.ts: readNatsAuthFlags hoisted out of readNatsRuntimeBoundary.
- background-tracing.ts: recordFailure/recordSuccess hoisted out of observe,
  both taking one ObservationContext so the parameter count stays within the
  rule. Tracing stays best-effort and off the job control path.
- auth.ts: rejectDelegationHeaders hoisted out of the middleware and the
  request augmentation typed via AuthenticatedRequest instead of any. The
  403/400 refusals fire in the same order for the same inputs.
- auth.test.ts, background-tracing.test.ts, nats-runtime.test.ts: oversized
  describe blocks split into sibling describes and the any/non-null casts
  replaced (invokeMiddleware, expectDefined). No test body, name, or
  assertion changed.

server/ importers repointed, including the type-only ones: server/application/nats.ts,
server/observability/{trace-sink,trace-types,langfuse-tracing}.ts,
server/tools/{types,operator-doctor,dependency-arrival.test}.ts.

Per M2 the three old src/ paths are now re-export shims, so src/ and
scripts/ importers keep their import lines untouched; the shims retire with
src/ at L6.

src/ runtime closure under server/main.ts: 67 -> 64.
